新增全局异常捕捉处理，添加vue代码模板

+++ simpleblogserver/src/main/java/com/chang/simpleblog/basedata/dao/IUserDao.java
@@ -13,7 +13,7 @@ import java.util.List;
  * @Version 1.0
  */
 @Repository
-public interface IUserDao extends JpaRepository<User,Long> {
+public interface IUserDao extends JpaRepository<User,Integer> {
 
     /**
      *  根据username查询用户信息
